requirements.txt
coloredlogs~=15.0.1
inflect~=5.3.0
ruamel.yaml~=0.17.19
pywin32==303; sys_platform == 'win32'
selenium~=4.1.0
selenium_stealth~=1.0.6
webdriver_manager~=3.5.2